Refactor taskle.model.person to taskle.model.task
