+++ src/main/java/taskle/logic/LogicManager.java
@@ -7,7 +7,7 @@ import taskle.logic.commands.Command;
 import taskle.logic.commands.CommandResult;
 import taskle.logic.parser.Parser;
 import taskle.model.Model;
-import taskle.model.person.ReadOnlyTask;
+import taskle.model.task.ReadOnlyTask;
 import taskle.storage.Storage;
 
 import java.util.logging.Logger;
