+++ src/main/java/taskle/logic/commands/EditCommand.java
@@ -2,13 +2,11 @@ package taskle.logic.commands;
 
 import taskle.commons.core.Messages;
 import taskle.commons.core.ModifiableObservableList;
-import taskle.commons.core.UnmodifiableObservableList;
 import taskle.commons.exceptions.IllegalValueException;
-import taskle.model.person.ModifiableTask;
-import taskle.model.person.Name;
-import taskle.model.person.ReadOnlyTask;
-import taskle.model.person.UniqueTaskList.DuplicateTaskException;
-import taskle.model.person.UniqueTaskList.TaskNotFoundException;
+import taskle.model.task.ModifiableTask;
+import taskle.model.task.Name;
+import taskle.model.task.UniqueTaskList.DuplicateTaskException;
+import taskle.model.task.UniqueTaskList.TaskNotFoundException;
 
 /**
  * Edits a task identified using it's last displayed index from the task manager.
