+++ src/main/java/taskle/model/task/ModifiableTask.java
@@ -1,4 +1,4 @@
-package taskle.model.person;
+package taskle.model.task;
 
 import taskle.model.tag.UniqueTagList;
