+++ src/main/java/taskle/storage/XmlSerializableTaskManager.java
@@ -5,10 +5,10 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 import taskle.commons.exceptions.IllegalValueException;
 import taskle.model.ReadOnlyTaskManager;
-import taskle.model.person.ReadOnlyTask;
-import taskle.model.person.UniqueTaskList;
 import taskle.model.tag.Tag;
 import taskle.model.tag.UniqueTagList;
+import taskle.model.task.ReadOnlyTask;
+import taskle.model.task.UniqueTaskList;
 
 import java.util.ArrayList;
 import java.util.Collections;
