+++ src/test/java/guitests/AddCommandTest.java
@@ -1,14 +1,14 @@
 package guitests;
 
-import guitests.guihandles.TaskCardHandle;
+import static org.junit.Assert.assertTrue;
+
 import org.junit.Test;
 
+import guitests.guihandles.TaskCardHandle;
 import taskle.commons.core.Messages;
 import taskle.logic.commands.AddCommand;
 import taskle.testutil.TestTask;
 import taskle.testutil.TestUtil;
-
-import static org.junit.Assert.assertTrue;
 
 public class AddCommandTest extends AddressBookGuiTest {
